+++ app/src/main/java/com/dudoji/android/config/MapConstConfig.kt
@@ -1,0 +1,17 @@
+package com.dudoji.android.config
+
+// location diff threshold for update
+const val LOCATION_UPDATE_THRESHOLD = 5 // meter
+
+const val REVEAL_CIRCLE_RADIUS_BY_WALK = 100.0 // meter
+
+// for swapping Tile overlay
+const val TILE_OVERLAY_LOADING_TIME = 700L // ms
+
+// for MapSection's Basic
+const val BASIC_ZOOM_LEVEL = 15
+
+
+// Google Map Config
+const val MIN_ZOOM = 10f
+const val MAX_ZOOM = 20f
