+++ app/src/main/java/com/dudoji/android/model/TileCoordinate.kt
@@ -1,0 +1,3 @@
+package com.dudoji.android.model
+
+data class TileCoordinate(val x: Int, val y: Int, val zoom: Int)
