+++ app/src/main/java/com/dudoji/android/util/location/LocationCallbackFilter.kt
@@ -1,10 +1,10 @@
 package com.dudoji.android.util.location
 
 import android.location.Location
+import com.dudoji.android.config.LOCATION_UPDATE_THRESHOLD
 
 object LocationCallbackFilter {
 
-    const val LOCATION_UPDATE_THRESHOLD = 5 // meter
     val lastLocation: Location = Location("")
 
     fun isSameLocation(location: Location): Boolean {
